requirements.txt
Flask==2.3.3
flask_sqlalchemy==3.0.5
flask_migrate==4.0.4
psycopg2-binary==2.9.7
gunicorn==21.2.0
alembic==1.12.0
pytest==7.4.3
pytest-cov==4.1.0
